+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Additions and Fixes
 
 * Fix definition of `on` in the standard library
+* Add `Map_keys` and `Map_fold` to the standard library.
 
 ## Version 1.0.4